* Common/MemoryUtil.h (getMemoryUsage): Portability fix.
git-svn-id: https://svn.bcgsc.ca/svn/abyss/trunk@73323 ad02020f-2a26-4133-947e-1829317c1416

+++ Common/MemoryUtil.h
@@ -14,7 +14,7 @@ static inline ssize_t getMemoryUsage()
 	std::ifstream in("/proc/self/statm");
 	size_t size, resident, share, text, lib, data;
 	return in >> size >> resident >> share >> text >> lib >> data
-		? data * getpagesize() : -1;
+		? ssize_t(data * getpagesize()) : -1;
 #else
 	/** Start of the data segment. */
 	static intptr_t sbrk0 = reinterpret_cast<intptr_t>(sbrk(0));
